requirements.txt
torch==1.13.1
torchvision==0.14.1
transformers==4.26.1
diffusers==0.10.2
PILLOW==9.0.1
matplotlib==3.6.2
opencv-python==4.6.0.66
numpy==1.23.3
